pyproject.toml
[tool.poetry]
name = "corpsystem-api"
version = "0.1.0"
description = ""
authors = ["jorgebarcelos <jorgebarcelos.faria@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.7"
django-ninja = "^1.0.1"
reportlab = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
